ques-practice

+++ 10kadane's Algo/max-subarray-sum.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-// #include <climits>
+#include <climits>
 using namespace std;
 
 // how to print subarray
@@ -27,7 +27,7 @@ using namespace std;
 int main() {
     int arr[7] = {3, -4, 5, 4, -1, 7, -8};
     int n = 7;
-    int maxSum = INT8_MIN;
+    int maxSum = INT_MIN;
 
     for (int st = 0; st < n; st++) {
         int currSum = 0;
@@ -48,7 +48,7 @@ int main() {
 //   int arr[7] = {3, -4, 5, 4, -1, 7, -8};
 //   int n = 7;
 //   int currSum = 0;
-//   int maxSum = INT8_MIN;
+//   int maxSum = INT_MIN;
 //   for (int i = 0; i < n; i++)
 //   {
 //     currSum += arr[i];
